[MOVE] /Global/Extension 내부 파일들 이동(#516)
