--- Manito/Manito/Global/Extension/UITableView+Extension.swift
@@ -1,22 +1,0 @@
-//
-//  UITableView+Extension.swift
-//  Manito
-//
-//  Created by SHIN YOON AH on 2022/06/09.
-//
-
-import UIKit
-
-extension UITableView {
-    func dequeueReusableCell<T: UITableViewCell>(withType cellType: T.Type, for indexPath: IndexPath) -> T {
-        guard let cell = self.dequeueReusableCell(withIdentifier: T.className, for: indexPath) as? T else {
-            fatalError("Could not find cell with reuseID \(T.className)")
-        }
-        return cell
-    }
-    
-    func register<T>(cell: T.Type,
-                      forCellReuseIdentifier reuseIdentifier: String = T.className) where T: UITableViewCell {
-        register(cell, forCellReuseIdentifier: reuseIdentifier)
-    }
-}
