+++ Manito/Manito/Util/Extension/Combine/Publisher+Async.swift
@@ -1,5 +1,5 @@
 //
-//  Publisher+Extension.swift
+//  Publisher+Async.swift
 //  Manito
 //
 //  Created by SHIN YOON AH on 2023/07/25.
