+++ Manito/Manito/Util/Extension/Type/Int+RandomNumber.swift
@@ -1,5 +1,5 @@
 //
-//  Int+Extension.swift
+//  Int+RandomNumber.swift
 //  Manito
 //
 //  Created by SHIN YOON AH on 2022/06/16.
