+++ Manito/Manito/Util/Extension/Type/NSObject+ClassName.swift
@@ -1,5 +1,5 @@
 //
-//  NSObject+Extension.swift
+//  NSObject+ClassName.swift
 //  Manito
 //
 //  Created by SHIN YOON AH on 2022/06/09.
