+++ Manito/Manito/Util/Extension/UI/UIViewController/UIViewController+Alert.swift
@@ -1,8 +1,8 @@
 //
-//  UIViewController+Extension.swift
+//  UIViewController+Alert.swift
 //  Manito
 //
-//  Created by SHIN YOON AH on 2022/06/09.
+//  Created by SHIN YOON AH on 2023/09/06.
 //
 
 import UIKit
@@ -17,10 +17,10 @@ extension UIViewController {
                                                     preferredStyle: .alert)
         let okAction = UIAlertAction(title: "확인", style: .default, handler: okAction)
         alertViewController.addAction(okAction)
-        
+
         self.present(alertViewController, animated: true, completion: completion)
     }
-    
+
     func makeRequestAlert(title: String,
                           message: String,
                           okTitle: String = "확인",
@@ -31,41 +31,31 @@ extension UIViewController {
                           completion : (() -> Void)? = nil) {
         let generator = UIImpactFeedbackGenerator(style: .medium)
         generator.impactOccurred()
-        
+
         let alertViewController = UIAlertController(title: title, message: message,
                                                     preferredStyle: .alert)
-        
+
         let cancelAction = UIAlertAction(title: cancelTitle, style: .default, handler: cancelAction)
         alertViewController.addAction(cancelAction)
-        
+
         let okAction = UIAlertAction(title: okTitle, style: okStyle, handler: okAction)
         alertViewController.addAction(okAction)
-        
+
         self.present(alertViewController, animated: true, completion: completion)
     }
-    
+
     func makeActionSheet(title: String? = nil,
                          message: String? = nil,
                          actionTitles:[String?],
                          actionStyle:[UIAlertAction.Style],
                          actions:[((UIAlertAction) -> Void)?]) {
         let alert = UIAlertController(title: title, message: message, preferredStyle: .actionSheet)
-        
+
         for (index, title) in actionTitles.enumerated() {
             let action = UIAlertAction(title: title, style: actionStyle[index], handler: actions[index])
             alert.addAction(action)
         }
-        
+
         self.present(alert, animated: true, completion: nil)
-    }
-    
-    func hidekeyboardWhenTappedAround() {
-        let tap = UITapGestureRecognizer(target: self, action: #selector(endEditingView))
-        tap.cancelsTouchesInView = false
-        view.addGestureRecognizer(tap)
-    }
-    
-    @objc func endEditingView() {
-        view.endEditing(true)
     }
 }
